moving check for cllection address in pallet
